chore(args): refactor
