+++ crates/release_plz/src/args/update.rs
@@ -1,27 +1,8 @@
 use std::path::PathBuf;
 
-use anyhow::{anyhow, Context};
+use anyhow::Context;
 use chrono::{Date, NaiveDate, Utc};
-use release_plz_core::{ChangelogRequest, GitHub, UpdateRequest, CARGO_TOML};
-use secrecy::SecretString;
-use url::Url;
-
-#[derive(clap::Parser, Debug)]
-#[clap(about, version, author)]
-pub struct CliArgs {
-    #[clap(subcommand)]
-    pub command: Command,
-}
-
-#[derive(clap::Subcommand, Debug)]
-pub enum Command {
-    /// Update packages version based on commit messages.
-    Update(Update),
-    /// Create a Pull Request representing the next release.
-    /// The Pull request contains updated packages version and changelog based on commit messages.
-    /// Close old PRs opened by release-plz, too.
-    ReleasePr(ReleasePr),
-}
+use release_plz_core::{ChangelogRequest, UpdateRequest, CARGO_TOML};
 
 #[derive(clap::Parser, Debug)]
 pub struct Update {
@@ -54,18 +35,6 @@ pub struct Update {
     /// If unspecified, crates.io is used.
     #[clap(long, conflicts_with("registry-project-manifest"))]
     registry: Option<String>,
-}
-
-#[derive(clap::Parser, Debug)]
-pub struct ReleasePr {
-    #[clap(flatten)]
-    pub update: Update,
-    /// GitHub token used to create the pull request.
-    #[clap(long)]
-    pub github_token: SecretString,
-    /// GitHub repository url where your project is hosted.
-    #[clap(long)]
-    pub repo_url: Url,
 }
 
 impl Update {
@@ -108,29 +77,5 @@ impl Update {
                 .expect("cannot retrieve current directory")
                 .join(CARGO_TOML),
         }
-    }
-}
-
-impl ReleasePr {
-    pub fn github(&self) -> anyhow::Result<GitHub> {
-        let segments = self
-            .repo_url
-            .path_segments()
-            .map(|c| c.collect::<Vec<_>>())
-            .ok_or_else(|| {
-                anyhow!(
-                    "cannot find github owner and repo from url {}",
-                    self.repo_url
-                )
-            })?;
-        let owner = segments
-            .get(0)
-            .ok_or_else(|| anyhow!("cannot find github owner from url {}", self.repo_url))?
-            .to_string();
-        let repo = segments
-            .get(1)
-            .ok_or_else(|| anyhow!("cannot find github repo from url {}", self.repo_url))?
-            .to_string();
-        Ok(GitHub::new(owner, repo, self.github_token.clone()))
     }
 }
